requirements.txt
torch
torchvision
albumentations
opencv-python
pillow
numpy
matplotlib
tqdm
